6736 ZFS per-vdev ZAPs
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: John Kennedy <john.kennedy@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

vdevs currently have a packed nvlist for config stored in their on-disk
labels. Instead of adding to this every time we want a new config, we
can make a new ZAP for every vdev and store its object # instead. It's
also confusing that the config is stored both in the label and in the
MOS, and some fields of the config are only valid in the MOS (the
version in the label could be out of date). These fields should be
stored in the new per-vdev ZAP objects.

Upstream-bugs: DLPX-38150 DLPX-39112