Properly handle updates of variably-sized SA entries
This change is a fusion of the following two changes:
- zfsonlinux/zfs@5d862cb
- freebsd/freebsd@a7cf665
plus some additional enhancements on top.

During the update process in sa_modify_attrs(), the sizes of existing
variably-sized SA entries are obtained from sa_lengths[]. The case where
a variably-sized SA was being replaced neglected to increment the index
into sa_lengths[], so subsequent variable-length SAs would be rewritten
with the wrong length. This patch adds the missing increment operation
so all variably-sized SA entries are stored with their correct lengths.

Another problem was that index into attr_desc[] was increased even when
an attribute was removed.  If that attribute was not the last attribute,
then the last attribute was lost.