6450 scrub/resilver unnecessarily traverses snapshots created after the scrub started Reviewed by: George Wilson <george.wilson@delphix.com> Reviewed by: Prakash Surya <prakash.surya@delphix.com>
If snapshots are being periodically created, scrub/resilver may never
complete (or may take a long time to complete). The symptom is that the
scrub gets "stuck" at 99+% done (according to "zpool status").

::zfs_dbgmsg reveals that we are traversing a snapshot that was created
long after the scrub started. In particular, the scn_cur_min_txg is >=
scn_cur_max_txg. This snapshot can't reference any blocks that need to
be scrubbed, so we are just wasting time reading its metadata looking
for blocks to scrub.