7501 Refactor spa_load_impl into several functions
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>

The work horse of the pool import process in the kernel is
spa_load_impl(). It is a very large function that is hard to track and
to debug. The plan here is to split it up into many well commented
functions so as to make the import process easier to follow.

The follow up to this cleanup process is to add debug messages for most
of the failure cases and then to create tools to facilitate the recovery
of a corrupted pool.

Upstream bugs: DLPX-42886