7586 remove #ifdef __lint hack from dmu.h
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>

The #ifdef __lint in dmu.h is ugly, and it would be nice not to
duplicate it if we add other inline functions into header files in ZFS,
especially since it is difficult to make any other solution work across
all compilation targets. We should switch to disabling the lint flags
that are failing instead.