Create tunable to ignore hole_birth feature
Until we can resolve the numerous hole_birth bugs that have cropped up recently, and come up with a way going forwards to protect users from corruption, we should disable the hole_birth feature.  Using a tunable allows those who are confident that their data is correct to continue to take advantage of the feature.