7200 7199 dsl_dataset_rollback_sync should not leave a dirty dataset
Sync-tasks that operate on existing datasets should not leave them
dirty, because there could be other sync-tasks for the same dataset
that do not expect to encounter a dirty dataset.
It's not a problem to leave a new dataset dirty as there can not be
any other sync-tasks for it.

Typically sync-tasks do not modify the dataset's objset as they
dirty only the MOS.  A single exception from this rule is when
a clone is created from a snapshot.  In that case we need to zero out
a ZIL header that could point to a ZIL chain that existed when
the snapshot was taken.
The only case when that exception is problematic is
dsl_dataset_rollback_sync() which creates a clone from a snapshot
and then performs the clone swap.

A sample of problems caused by the dirty datasets:
- if a dirty dataset is snapshotted, then there will be blocks that
  are not reachable from the root BP, but with birth txg equal to
  the snapshot txg
- if a dirty dataset is destroyed, then there will be a crash when
  attempting to sync out the dataset
- if a rolled back dataset is rolled back again, then the dirty
  intermediate dataset may have a non-zero ZIL header pointing to
  a ZIL chain captured by the snapshot, when the intermediate dataset
  is destroyed the stale chain would be walked