7572 vioif panic: qe->qe_indirect_next < qe->qe_queue->vq_indirect_num 7574 boot slowness followed by panic while booting on KVM when no cpu tag is specified in virsh XML
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Prachetaa Raghavan <prachetaa.raghavan@delphix.com>
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Robert Mustacchi <rm@joyent.com>
Reviewed by: Igor Kozhukhov <ikozhukhov@gmail.com>

These are 2 fixes for running illumos on KVM:

The first is a panic where we access beyond the end of an array because
the calculation in virtio_ve_indirect_available() is off by one.

The second handles the case where KVM is not provided with a processor
tag in the XML file used to describe the VM to libvirt. Today this
results in the guest misunderstanding the processor type, which causes
the VM to be very slow. The fix recognizes this case based on the
processor family / model numbers and the fact that we're running
inside KVM, and compensates for the issue by incrementing xcpuid.