7569 statd support to run on a fixed port 7577 mountd support to run on a fixed port
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: Matt Amdur <matt.amdur@delphix.com>
Reviewed by: John Kennedy <john.kennedy@delphix.com>

A documented list of fixed network ports is needed when configuring
firewall and IDS bypass rules. statd uses a random port which makes it
difficult to configure the firewall rules. Some firewalls cannot
dynamically bypass a port-mapper allocated port number. This provides a
configuration option to run statd and mountd on a fixed port.