9280 Assertion failure while running removal_with_ganging test with 4K devices
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: John Kennedy <john.kennedy@delphix.com>

We hit a panic in the removal_with_ganging test using vdevs with 4k ashift for testing.

The problem is that the test intentionally induces ganging in allocations as small as 4K, but failing a minimum-size allocation (4K in this case) is not permitted. We should change the test to only gang 8K and larger, or make it dynamic based on the max ashift in the pool.

Note: the operative fix is removal_with_ganging.ksh line 37, the rest is renaming.

Upstream bug: DLPX-53029